+++ pages/ingredients.php
@@ -160,7 +160,10 @@ $res_ingCategory = mysqli_query($conn, "SELECT id,image,name,notes FROM ingCateg
       <strong>WARNING:</strong><br />
       you are about to import data from PV Online, please bear in mind, PV Online is a community driven database therefore may contain unvalidated or incorrect data. <br />
       If your local database contains already an ingredient with the same name, the ingredient data will not be imported. <p></p>
-      Ingredients online: <strong><?php echo pvOnlineStats($pvOnlineAPI, 'ingredients');?></strong>
+      <p>Ingredients online: <strong><?php echo pvOnlineStats($pvOnlineAPI, 'ingredientsTotal');?></strong></p>
+      <p>Synonyms online: <strong><?php echo pvOnlineStats($pvOnlineAPI, 'synonymsTotal');?></strong></p>
+      <p>Compositions online: <strong><?php echo pvOnlineStats($pvOnlineAPI, 'composTotal');?></strong></p>
+
 </div>
 	  <div class="modal-footer_2">
 	  <?php require('privacy_note.php');?>