+++ releasenotes.md
@@ -3,3 +3,5 @@ Whats New:
 - Improved installtion wizard 
 - Pending formulas moved to Formula Management
 - New notification system implemented
+- Update to comply with new PV Online api
+- Show ingredients, synonyms and compositions count when you importing from PV Online